Use local instance of ZipArchive
As there seems to be no way to flush a ZipArchive to disk reliably while keeping
it open, just use a local ZipArchive that gets closed/flushed instantly.
Ideally, batching write events and/or using a completely different
mechanism, but considering the request rate that this backend will have,
all other optimizations seem to be premature.